api/types/system: remove Info.ExecutionDriver
The execution-driver was replaced with containerd since docker 1.11 (API
v1.23) in 9c4570a958df42d1ad19364b1a8da55b891d850a, after which the value
was no longer set. The field was left in the type definition.
Commit 1fb1136fecfd761300a38f64ac9178979cc0b270 removed its use from the
CLI and [docker/engine-api@39c7d7e] removed it from the API type, followed
by an update to the API docs in 3c6ef4c29d28e92ea29816d6117412162d829c60.

Changes to the API types were not pulled into the engine until v1.13, and
probably because of that gated it on API version < 1.25 instead of < 1.24
(see 6d98e344c7702a8a713cb9e02a19d83a79d3f930); setting a "not supported"
value for older versions.

Based on the above; this field was deprecated in API v1.23, and empty
since then. Given that the minimum API version supported by the engine
is not v1.24, we can safely remove it.

[docker/engine-api@39c7d7e]: https://github.com/docker/engine-api/commit/39c7d7ec192b065f757ffdb85b36a7b3ede4e10b

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>