libcontainerd/supervisor: remove remnants of adjusting oom-score
In docker versions before v24.0, the dockerd daemon allowed adjusting
its own oom-score; in that case we would forward the configuration to
containerd (when running as child process) for it to do the same.

This feature was deprecated in 5a922dc162bbe0a03450165da4e6aceca55073d4,
and removed in fb96b94ed00aa4f200dc03642bc46d4289eb6860, but some fields
were left behind in the supervisor package.

This patch removes the oomScore field and adjustOOMScore method from
the remove type.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>