cmd/dockerd: windows: don't use Fatal log to prevent early exit
cmd/dockerd: notifyShutdown: don't use Fatal log to prevent early exit
------------------------------------------------------------------------

Commit 40868e263d7a239aa10acd922bd69ba5a943781e added this log for situations
where the daemon failed to start, and where no event-log was created
to discover the failure.

However, it used a `Fatal` log which, when using logrus, not only logs
a message, but also calls the [logger's Exit function][1], which defaults
[to `os.Exit`][2].

The error passed to the `notifyShutdown` function can be the result of
different steps in the service's lifecycle; `cli.start()` starts a long-
lived process, and the error returned can be either because it failed
to start the service, or an error produced while shutting down the service.

We should reduce ambiguity here (which could be for `cli.Start()` to return
whether the service was started in the first place). Regardless, we want
code following the log-entry to be executed, to make sure that codepaths
following the `notifyShutdown` (which could be (defer) statements or cleanup
steps) are executed.

This patch changes the log to an `Error`, which is non-fatal, and should
make sure that at least the `service.stopped` is executed to signal the
service no longer running. The logging is also moved out of the
`notifyShutdown`.

Disclaimer: this is with the assumption that calling `service.stopped` is
always acceptable, even if the service is not running, or in other states.

cmd/dockerd: preNotifyReady: return error instead of logging and exit
------------------------------------------------------------------------

Commit 57aef3b49025aac6bb084491478eb461b14b9109 added support for running
the daemon as a Windows service. As part of this, it introduced a log
for situations where the service failed to start. However, it used a
`Fatal` log for this purpose which, when using logrus, not only logs
a message, but also calls the [logger's Exit function][1], which defaults
[to `os.Exit`][2].

A result of this is that the daemon does an unclean shutdown, causing
other code (which could be (defer) statements or cleanup steps) not to
be executed.

This patch updates the `preNotifyReady` to return the error instead of
logging and exiting. `preNotifyReady` is executed as part of `cli.start()`,
which has an error-return, and the error that is returned, is already
logged by `runDaemon` (so logging it as part of `preNotifyReady` would
result in the error being logged twice).

[1]: https://github.com/moby/moby/blob/5098132d84d3c20afd05453e430b5029e770fcc1/vendor/github.com/sirupsen/logrus/entry.go#L336-L339
[2]: https://github.com/moby/moby/blob/5098132d84d3c20afd05453e430b5029e770fcc1/vendor/github.com/sirupsen/logrus/logger.go#L342-L348

- relates to https://github.com/moby/moby/pull/22340

cmd/dockerd: preNotifyReady: return error instead of logging and exit

Commit 57aef3b49025aac6bb084491478eb461b14b9109 added support for running
the daemon as a Windows service. As part of this, it introduced a log
for situations where the service failed to start. However, it used a
`Fatal` log for this purpose which, when using logrus, not only logs
a message, but also calls the [logger's Exit function][1], which defaults
[to `os.Exit`][2].

A result of this is that the daemon does an unclean shutdown, causing
other code (which could be (defer) statements or cleanup steps) not to
be executed.

This patch changes the `preNotifyReady` to

[1]: https://github.com/moby/moby/blob/5098132d84d3c20afd05453e430b5029e770fcc1/vendor/github.com/sirupsen/logrus/entry.go#L336-L339
[2]: https://github.com/moby/moby/blob/5098132d84d3c20afd05453e430b5029e770fcc1/vendor/github.com/sirupsen/logrus/logger.go#L342-L348

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>