daemon: remove configsSupported utility
This utility was originally added in 9e9fc7b57c1764c008e568ed52bcd1aade7eb40c
at which time it was not yet implemented for Windows, so this utility was
used to print a warning when trying to use it on that platform.

Windows support was added in e0d533b1e8a8dd62ed6dff2dfda3c3220e0474b9, which
kept the utility, but adjusted it to support for both Windows and Linux, and
excluding any other platform.

Let's remove this utility, given that we currently only support Windows and
Linux (there's been some partial support for other platforms, but they are
very likely broken in many ways).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>